jobcard update
