+++ tsfrontend/src/app/App.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import Router from "app/Router";
 import { Provider } from "react-redux";
 import store from "redux/store";
+import "app/app.css";
 
 export default function App() {
   return (
