+++ tsfrontend/src/app/app.css
@@ -1,0 +1,5 @@
+/* @media (width: 999px) {
+  .card-columns {
+    column-count: 2;
+  }
+} */
