+++ tsfrontend/src/jobs/Search.tsx
@@ -45,7 +45,7 @@ const Search: FC<RouteComponentProps<any>> = (props) => {
     <Container<React.ElementType> className="pt-2 pb-2" align="center">
       <Form onSubmit={(e: React.FormEvent<HTMLFormElement>) => handleSubmit(e)}>
         <Form.Row style={{ justifyContent: "space-between" }}>
-          <Col xs="10">
+          <Col xs={6} sm={7} lg={10}>
             <TextField
               type="text"
               name="search"
